Update to python 3.7 and latest nuklear
Everything except the overview translated to python seems to work
perfectly. The biggest change between versions was that now with
every string on cffi you have to do "test".encode('utf-8') instead
of just using a string in a cffi function like "test". I have not
found a way to mitigate this issue so far. This change in encoding
practice with cffi is the source of almost all braking in
overview.py. I used a vim regex command to add encode('utf-8') to
all strings that aren't doc strings, but there are several issues
still present and undiscovered, and can only be discovered by
running overview and pressing on random buttons.